Refactor user views: add OwnerRequiredMixin and get_absolute_url
