+++ backend/users/models.py
@@ -1,3 +1,4 @@
+from django.urls import reverse
 from django.contrib.auth.models import AbstractUser
 from django.contrib.auth.validators import UnicodeUsernameValidator
 from django.core.exceptions import ValidationError
@@ -59,3 +60,6 @@ class User(AbstractUser):
 
     def __str__(self):
         return self.username
+
+    def get_absolute_url(self):
+        return reverse('users:profile', kwargs={'username': self.username})
